refactored addcoursetopath requires styling fixes
