+++ didact/src/components/courses/Course.jsx
@@ -8,8 +8,7 @@ import {
   postCourseToPath
 } from "../../store/actions/index";
 
-import { AddCourseToPath, PopoverWrapper } from "./CourseStyles";
-import { DidactButton } from "../dashboard/ButtonStyles";
+import { PopoverWrapper } from "./CourseStyles";
 
 import { makeStyles } from "@material-ui/core/styles";
 import ArrowRightAltRoundedIcon from "@material-ui/icons/ArrowRightAltRounded";
@@ -17,11 +16,9 @@ import Card from "@material-ui/core/Card";
 import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
-import Popover from "@material-ui/core/Popover";
 import ExpandLessIcon from "@material-ui/icons/ExpandLess";
 
-import playlistAdd from "../../images/playlist_add_black_24x24.png";
-import closeIcon from "../../images/close_black_24x24.png";
+import AddCoursePathPlaylist from "./AddCoursePathPlaylist";
 
 const useStyles = makeStyles(theme => ({
   buttonCourse: {
@@ -94,16 +91,6 @@ const useStyles = makeStyles(theme => ({
   title: {
     display: "flex"
   },
-  addCourse: {
-    background: "none",
-    border: "black",
-    height: "100%",
-    display: "flex",
-    justifyContent: "flex-end",
-    margin: "-21px -4px 15px 0",
-    position: "relative",
-    zIndex: 12
-  },
   courseTitle: {
     maxWidth: "512px",
     color: "black"
@@ -126,45 +113,19 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-const Course = ({ course, addingCourses, props, tracked }) => {
+const Course = ({ course, props, tracked }) => {
   const state = useSelector(state => state);
   const classes = useStyles();
   const dispatch = useDispatch();
-  const learningPaths = state.learningPathReducer.yourLearningPathsOwned;
-  const filteredPaths = [];
   const [expanded, setExpanded] = useState(false);
-  const [anchorEl, setAnchorEl] = useState(null);
-  const open = Boolean(anchorEl);
-  const id = open ? "simple-popover" : undefined;
 
-  // useEffect(() => {
-  //   dispatch(courseEndPoint());
-  // }, [dispatch]);
-
-  // useEffect(() => {
-  //   dispatch(getYourLearningPathsOwned());
-  // }, [dispatch, state.learningPathReducer.learningPath]);
+  useEffect(() => {
+    dispatch(getYourLearningPathsOwned());
+  }, [dispatch, state.learningPathReducer.learningPath]);
 
   const handleExpandClick = () => {
     setExpanded(!expanded);
   };
-
-  const handleClick = event => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const handleAddCourse = (path_id, course_id, order) => {
-    dispatch(postCourseToPath(path_id, course_id, Number(order)));
-    setAnchorEl(null);
-  };
-
-  learningPaths.forEach(path => {
-    if (!path.courseIds.includes(course.id)) filteredPaths.push(path);
-  });
 
   return (
     <PopoverWrapper>
@@ -177,87 +138,7 @@ const Course = ({ course, addingCourses, props, tracked }) => {
         }
       >
         <CardContent>
-          <div
-            style={{
-              display: "flex",
-              justifyContent: "flex-end",
-              paddingTop: "20px"
-            }}
-          >
-            {addingCourses && (
-              <button className={classes.addCourse} onClick={handleClick}>
-                <img src={playlistAdd} alt="Add Course" />
-              </button>
-            )}
-          </div>
-          <div>
-            <Popover
-              id={id}
-              open={open}
-              anchorEl={anchorEl}
-              onClose={handleClose}
-              anchorOrigin={{
-                vertical: "bottom",
-                horizontal: "right"
-              }}
-              transformOrigin={{
-                vertical: "center",
-                horizontal: "right"
-              }}
-            >
-              {
-                <AddCourseToPath>
-                  {
-                    <div>
-                      <div
-                        style={{ marginTop: "10px", paddingRight: "5px" }}
-                        className="closePopover"
-                      >
-                        <img
-                          src={closeIcon}
-                          onClick={handleClose}
-                          alt="Close"
-                        />
-                      </div>
-                      <div className="learningPaths">
-                        <h4 style={{ margin: " -5px auto" }}>
-                          Add to Learning Path
-                        </h4>
-                        {filteredPaths.length > 0 ? (
-                          filteredPaths.length > 0 &&
-                          filteredPaths.map((learningPath, index) => {
-                            return (
-                              <div className="learningPathTitle" key={index}>
-                                <h5>{learningPath.title}</h5>
-                                <button
-                                  onClick={() =>
-                                    handleAddCourse(
-                                      learningPath.id,
-                                      course.id,
-                                      learningPath.contentLength + 1
-                                    )
-                                  }
-                                >
-                                  <img src={playlistAdd} alt="Add Course" />
-                                </button>
-                              </div>
-                            );
-                          })
-                        ) : (
-                          <p>Can't Add Course To Any Learning Paths</p>
-                        )}
-                      </div>
-                      <div className="buttons">
-                        <DidactButton onClick={handleClose}>Done</DidactButton>
-                        <a href="/learning-paths/add">Create Learning Path</a>
-                      </div>
-                    </div>
-                  }
-                </AddCourseToPath>
-              }
-            </Popover>
-          </div>
-
+          <AddCoursePathPlaylist course={course} />
           <CardActions disableSpacing>
             <div
               style={{
